moved layout to layouts folder and renamed properly
